[FFI][REFACTOR] Cleanup API locations
This PR cleans up the env api and move it to extra.
