+++ apps/android_rpc/app/src/main/jni/tvm_runtime.h
@@ -41,10 +41,10 @@
 #include "../ffi/src/ffi/extra/library_module_dynamic_lib.cc"
 #include "../ffi/src/ffi/extra/library_module_system_lib.cc"
 #include "../ffi/src/ffi/extra/module.cc"
+#include "../ffi/src/ffi/extra/testing.cc"
 #include "../ffi/src/ffi/function.cc"
 #include "../ffi/src/ffi/ndarray.cc"
 #include "../ffi/src/ffi/object.cc"
-#include "../ffi/src/ffi/testing.cc"
 #include "../ffi/src/ffi/traceback.cc"
 #include "../src/runtime/cpu_device_api.cc"
 #include "../src/runtime/device_api.cc"
