+++ ffi/src/ffi/extra/library_module_system_lib.cc
@@ -123,7 +123,7 @@ TVM_FFI_STATIC_INIT_BLOCK({
 }  // namespace ffi
 }  // namespace tvm
 
-int TVMFFIEnvRegisterSystemLibSymbol(const char* name, void* ptr) {
+int TVMFFIEnvModRegisterSystemLibSymbol(const char* name, void* ptr) {
   tvm::ffi::SystemLibSymbolRegistry::Global()->RegisterSymbol(name, ptr);
   return 0;
 }
