+++ ffi/src/ffi/extra/module.cc
@@ -130,8 +130,8 @@ TVM_FFI_STATIC_INIT_BLOCK({
 }  // namespace ffi
 }  // namespace tvm
 
-int TVMFFIEnvLookupFromImports(TVMFFIObjectHandle library_ctx, const char* func_name,
-                               TVMFFIObjectHandle* out) {
+int TVMFFIEnvModLookupFromImports(TVMFFIObjectHandle library_ctx, const char* func_name,
+                                  TVMFFIObjectHandle* out) {
   TVM_FFI_SAFE_CALL_BEGIN();
   *out = tvm::ffi::ModuleObj::InternalUnsafe::GetFunctionFromImports(
       reinterpret_cast<tvm::ffi::ModuleObj*>(library_ctx), func_name);
