+++ web/emcc/wasm_runtime.cc
@@ -54,10 +54,10 @@
 #include "ffi/src/ffi/extra/library_module.cc"
 #include "ffi/src/ffi/extra/library_module_system_lib.cc"
 #include "ffi/src/ffi/extra/module.cc"
+#include "ffi/src/ffi/extra/testing.cc"
 #include "ffi/src/ffi/function.cc"
 #include "ffi/src/ffi/ndarray.cc"
 #include "ffi/src/ffi/object.cc"
-#include "ffi/src/ffi/testing.cc"
 #include "ffi/src/ffi/traceback.cc"
 #include "src/runtime/memory/memory_manager.cc"
 #include "src/runtime/nvtx.cc"
